Splitting main part of app out of the main bundle.
Also improving the transition from intro to compressor.
