+++ src/codecs/imagequant/options.tsx
@@ -1,5 +1,6 @@
 import { h, Component } from 'preact';
-import { bind, inputFieldValueAsNumber, konami } from '../../lib/util';
+import { bind } from '../../lib/initial-util';
+import { inputFieldValueAsNumber, konami } from '../../lib/util';
 import { QuantizeOptions } from './quantizer';
 
 const konamiPromise = konami();
