+++ src/codecs/mozjpeg/options.tsx
@@ -1,5 +1,6 @@
 import { h, Component } from 'preact';
-import { bind, inputFieldChecked, inputFieldValueAsNumber } from '../../lib/util';
+import { bind } from '../../lib/initial-util';
+import { inputFieldChecked, inputFieldValueAsNumber } from '../../lib/util';
 import { EncodeOptions, MozJpegColorSpace } from './encoder';
 import '../../custom-els/RangeInput';
